fix(clippy): ADR-0077 Phase A+B - promote correctness/perf lints to warn
Phase A: Cargo.toml lint promotions
- float_cmp: warn (exact float comparisons are correctness signals)
- significant_drop_tightening: warn (lock scope bugs)
- cast_precision_loss/truncation: warn (numeric overflow risks)
- redundant_clone: warn (performance optimization)
- missing_const_for_fn: warn (free const perf)
- missing_errors_doc/must_use_candidate: allow (Phase C deferred)

Phase B: Source-level #[allow] for justified cases
- Added module-level cast allows for math-heavy modules
  (hyperdim, reservoir_sparse, bm25, bridge_retrieval, etc.)
- Added function-level allows for justified lock scopes
  (probe, process_sequence, batch ops)
- Added test module allows for exact float comparisons
- Added redundant_clone allow for WASM ownership semantics

LOC compliance: Extracted modules to stay under 500
- framework.rs: 519 → 453 (extracted FrameworkMetrics to framework_metrics.rs)
- framework_ops.rs: 507 → 421 (extracted validate_path to framework_validation.rs)
- wasm_ext.rs: 505 → 468 (compacted test section)
- singularity.rs: 503 → 500 (compacted similar_empty test)
- bm25.rs: 502 → 499 (compacted test_clear)

All validation gates pass: clippy -D warnings, tests, LOC < 500
