+++ Cargo.toml
@@ -108,6 +108,26 @@ all = { level = "warn", priority = -1 }
 pedantic = { level = "allow", priority = -1 }
 nursery = { level = "allow", priority = -1 }
 
+# ============================================================================
+# ADR-0077: Promoted from pedantic — correctness signal (Phase A)
+# ============================================================================
+float_cmp = "warn"                      # Strict f32/f64 == almost always a bug
+significant_drop_tightening = "warn"    # Lock held longer than needed
+cast_precision_loss = "warn"            # usize as f32 losing bits
+cast_possible_truncation = "warn"       # i64 as u32 truncation risk
+redundant_clone = "warn"                # Free perf wins
+
+# ============================================================================
+# ADR-0077: Promoted from pedantic — library DX (Phase A allow, Phase C warn)
+# ============================================================================
+missing_errors_doc = "allow"            # Phase C: flip to warn after docs added
+must_use_candidate = "allow"            # Phase C: flip to warn after audit
+
+# ============================================================================
+# ADR-0077: Promoted from nursery — free perf wins
+# ============================================================================
+missing_const_for_fn = "warn"           # Const fn free perf
+
 # Explicit overrides for justified allows in source
 too_many_arguments = "allow"  # Internal hot paths documented in GOAP_STATE
 type_complexity = "allow"     # Complex generics in framework
