+++ llms.txt
@@ -205,9 +205,6 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 ### src/framework.rs
 
 - pub struct ChaoticSemanticFramework
-- pub struct FrameworkMetrics
-- pub struct FrameworkMetricsSnapshot
-- impl FrameworkMetrics
 - impl ChaoticSemanticFramework
 
 ### src/framework_bridge.rs
@@ -226,6 +223,12 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 
 - pub enum MemoryEvent
 - impl ChaoticSemanticFramework
+
+### src/framework_metrics.rs
+
+- pub struct FrameworkMetrics
+- pub struct FrameworkMetricsSnapshot
+- impl FrameworkMetrics
 
 ### src/framework_ops.rs
 
@@ -1095,6 +1098,26 @@ all = { level = "warn", priority = -1 }
 # Pedantic lints are too strict for this project
 pedantic = { level = "allow", priority = -1 }
 nursery = { level = "allow", priority = -1 }
+
+# ============================================================================
+# ADR-0077: Promoted from pedantic — correctness signal (Phase A)
+# ============================================================================
+float_cmp = "warn"                      # Strict f32/f64 == almost always a bug
+significant_drop_tightening = "warn"    # Lock held longer than needed
+cast_precision_loss = "warn"            # usize as f32 losing bits
+cast_possible_truncation = "warn"       # i64 as u32 truncation risk
+redundant_clone = "warn"                # Free perf wins
+
+# ============================================================================
+# ADR-0077: Promoted from pedantic — library DX (Phase A allow, Phase C warn)
+# ============================================================================
+missing_errors_doc = "allow"            # Phase C: flip to warn after docs added
+must_use_candidate = "allow"            # Phase C: flip to warn after audit
+
+# ============================================================================
+# ADR-0077: Promoted from nursery — free perf wins
+# ============================================================================
+missing_const_for_fn = "warn"           # Const fn free perf
 
 # Explicit overrides for justified allows in source
 too_many_arguments = "allow"  # Internal hot paths documented in GOAP_STATE
