+++ src/bridge_persistence.rs
@@ -3,6 +3,9 @@
 //! Feature-gated persistence layer for storing the symbolic semantic graph
 //! used by the bridge retrieval pipeline.
 
+// Casts are intentional for version serialization
+#![allow(clippy::cast_possible_truncation)]
+
 use crate::error::{MemoryError, Result};
 use crate::persistence::Persistence;
 use crate::semantic_bridge::{CanonicalConcept, ConceptGraph};
