+++ src/cli/commands/associate.rs
@@ -1,3 +1,8 @@
+//! Association commands for linking concepts.
+
+// Casts are intentional for CLI output formatting
+#![allow(clippy::cast_precision_loss, clippy::cast_possible_truncation)]
+
 use std::path::Path;
 
 use tracing::instrument;
