+++ src/cli/commands/probe.rs
@@ -1,3 +1,8 @@
+//! Probe commands for similarity search.
+
+// Casts are intentional for CLI output formatting
+#![allow(clippy::cast_precision_loss, clippy::cast_possible_truncation)]
+
 use std::path::Path;
 
 use tracing::instrument;
