+++ src/framework.rs
@@ -1,13 +1,13 @@
 //! Main framework integrating all components
 
 use std::sync::Arc;
-use std::sync::atomic::{AtomicU64, Ordering};
 use tokio::sync::RwLock;
 use tracing::{instrument, warn};
 
 use crate::error::Result;
 use crate::framework_builder::{FrameworkBuilder, FrameworkConfig, FrameworkStats};
 use crate::framework_events::MemoryEvent;
+use crate::framework_metrics::{FrameworkMetrics, FrameworkMetricsSnapshot};
 use crate::graph_traversal::TraversalConfig;
 use crate::hyperdim::HVec10240;
 use crate::metadata_filter::MetadataFilter;
@@ -27,71 +27,6 @@ pub struct ChaoticSemanticFramework {
     pub(crate) config: FrameworkConfig,
     pub(crate) metrics: Arc<FrameworkMetrics>,
     pub(crate) event_sender: tokio::sync::broadcast::Sender<MemoryEvent>,
-}
-
-#[derive(Debug, Default)]
-pub struct FrameworkMetrics {
-    concepts_injected_total: AtomicU64,
-    associations_created_total: AtomicU64,
-    probes_total: AtomicU64,
-    probe_latency_ms_total: AtomicU64,
-    probe_latency_count: AtomicU64,
-}
-
-#[derive(Debug, Clone)]
-pub struct FrameworkMetricsSnapshot {
-    pub concepts_injected_total: u64,
-    pub associations_created_total: u64,
-    pub probes_total: u64,
-    pub avg_probe_latency_ms: f64,
-    pub cache_hits_total: u64,
-    pub cache_misses_total: u64,
-    pub cache_evictions_total: u64,
-    pub reservoir_steps_total: u64,
-    pub avg_reservoir_step_latency_us: f64,
-    pub reservoir_nodes_active: u64,
-}
-
-impl FrameworkMetrics {
-    pub(crate) fn inc_concepts_injected(&self, count: u64) {
-        self.concepts_injected_total
-            .fetch_add(count, Ordering::Relaxed);
-    }
-
-    pub(crate) fn inc_associations_created(&self, count: u64) {
-        self.associations_created_total
-            .fetch_add(count, Ordering::Relaxed);
-    }
-
-    fn observe_probe_latency_ms(&self, latency_ms: u64) {
-        self.probes_total.fetch_add(1, Ordering::Relaxed);
-        self.probe_latency_ms_total
-            .fetch_add(latency_ms, Ordering::Relaxed);
-        self.probe_latency_count.fetch_add(1, Ordering::Relaxed);
-    }
-
-    fn snapshot(&self) -> FrameworkMetricsSnapshot {
-        let count = self.probe_latency_count.load(Ordering::Relaxed);
-        let total = self.probe_latency_ms_total.load(Ordering::Relaxed);
-        let avg = if count == 0 {
-            0.0
-        } else {
-            total as f64 / count as f64
-        };
-
-        FrameworkMetricsSnapshot {
-            concepts_injected_total: self.concepts_injected_total.load(Ordering::Relaxed),
-            associations_created_total: self.associations_created_total.load(Ordering::Relaxed),
-            probes_total: self.probes_total.load(Ordering::Relaxed),
-            avg_probe_latency_ms: avg,
-            cache_hits_total: 0,
-            cache_misses_total: 0,
-            cache_evictions_total: 0,
-            reservoir_steps_total: 0,
-            avg_reservoir_step_latency_us: 0.0,
-            reservoir_nodes_active: 0,
-        }
-    }
 }
 
 impl ChaoticSemanticFramework {
@@ -168,27 +103,42 @@ impl ChaoticSemanticFramework {
     }
 
     /// Query for similar concepts
+    #[allow(clippy::significant_drop_tightening)] // Lock needed for expired concept filtering
     #[instrument(err, skip(self, query))]
     pub async fn probe(&self, query: HVec10240, top_k: usize) -> Result<Vec<(String, f32)>> {
         self.validate_top_k(top_k)?;
         #[cfg(not(target_arch = "wasm32"))]
         let start = std::time::Instant::now();
-        let sing = self.singularity.read().await;
-        let results = sing.find_similar(&query, top_k);
+
+        // Acquire lock, get results, release immediately
+        let (results, expired_ids) = {
+            let sing = self.singularity.read().await;
+            let results = sing.find_similar(&query, top_k);
+
+            // Collect expired IDs while holding lock
+            let now = crate::singularity::unix_now_secs();
+            let expired_ids: std::collections::HashSet<String> = results
+                .iter()
+                .filter_map(|(id, _)| {
+                    sing.get(id)
+                        .and_then(|c| c.expires_at.filter(|exp| *exp <= now))
+                        .map(|_| id.clone())
+                })
+                .collect();
+            (results, expired_ids)
+        };
+
         #[cfg(not(target_arch = "wasm32"))]
+        #[allow(clippy::cast_possible_truncation)] // Duration millis to u64 for metrics
         let elapsed_ms = start.elapsed().as_millis() as u64;
         #[cfg(target_arch = "wasm32")]
         let elapsed_ms = 0;
         self.metrics.observe_probe_latency_ms(elapsed_ms);
 
-        // Filter expired concepts
-        let now = crate::singularity::unix_now_secs();
+        // Filter expired concepts without lock
         let filtered: Vec<(String, f32)> = results
             .into_iter()
-            .filter(|(id, _)| {
-                sing.get(id)
-                    .is_none_or(|c| c.expires_at.is_none_or(|exp| exp > now))
-            })
+            .filter(|(id, _)| !expired_ids.contains(id))
             .collect();
 
         Ok(filtered)
@@ -206,9 +156,15 @@ impl ChaoticSemanticFramework {
         Self::validate_metadata_filter(filter)?;
         #[cfg(not(target_arch = "wasm32"))]
         let start = std::time::Instant::now();
-        let sing = self.singularity.read().await;
-        let results = sing.find_similar_filtered(query, top_k, filter);
+
+        // Acquire lock, get results, release immediately
+        let results = {
+            let sing = self.singularity.read().await;
+            sing.find_similar_filtered(query, top_k, filter)
+        };
+
         #[cfg(not(target_arch = "wasm32"))]
+        #[allow(clippy::cast_possible_truncation)] // Duration millis to u64 for metrics
         let elapsed_ms = start.elapsed().as_millis() as u64;
         #[cfg(target_arch = "wasm32")]
         let elapsed_ms = 0;
@@ -239,6 +195,7 @@ impl ChaoticSemanticFramework {
     }
 
     /// Process temporal sequence through reservoir
+    #[allow(clippy::significant_drop_tightening)] // Reservoir lock needed for sequence processing
     #[instrument(err, skip(self, sequence))]
     pub async fn process_sequence(&self, sequence: &[Vec<f32>]) -> Result<HVec10240> {
         self.validate_sequence_length(sequence.len())?;
@@ -476,8 +433,11 @@ impl ChaoticSemanticFramework {
 
     /// Get framework statistics
     pub async fn stats(&self) -> Result<FrameworkStats> {
-        let sing = self.singularity.read().await;
-        let concept_count = sing.len();
+        // Get concept count without holding lock during persistence call
+        let concept_count = {
+            let sing = self.singularity.read().await;
+            sing.len()
+        };
 
         let db_size = if let Some(ref persistence) = self.persistence {
             Some(persistence.size().await.unwrap_or(0))
