+++ src/framework_events.rs
@@ -1,3 +1,8 @@
+//! Framework event broadcasting.
+
+// Casts are intentional for event timestamp math
+#![allow(clippy::cast_precision_loss, clippy::cast_possible_truncation)]
+
 use tokio::sync::broadcast;
 
 #[cfg(target_arch = "wasm32")]
@@ -108,6 +113,8 @@ pub(crate) fn build_event_sender() -> broadcast::Sender<MemoryEvent> {
 
 #[cfg(test)]
 mod tests {
+    #![allow(clippy::redundant_clone)] // Clone needed for event ownership testing
+
     use super::*;
 
     #[test]
