+++ src/singularity_retrieval.rs
@@ -6,6 +6,9 @@
 //! - `RetrievalConfig`: Configuration for candidate generation
 //! - Extension trait for reduced-candidate retrieval
 
+// Casts are intentional for retrieval similarity math
+#![allow(clippy::cast_precision_loss, clippy::cast_possible_truncation)]
+
 use serde::{Deserialize, Serialize};
 use std::collections::VecDeque;
 use std::sync::Arc;
@@ -103,7 +106,7 @@ impl Singularity {
     }
 
     /// Get the retrieval configuration.
-    pub fn retrieval_config(&self) -> &RetrievalConfig {
+    pub const fn retrieval_config(&self) -> &RetrievalConfig {
         &self.retrieval_config
     }
