+++ tests/api_completeness.rs
@@ -1,5 +1,8 @@
 //! Tests for Phase 33 API completeness features.
 
+// Locks held during test assertions for framework state verification
+#![allow(clippy::significant_drop_tightening)]
+
 use chaotic_semantic_memory::{ChaoticSemanticFramework, HVec10240, error::MemoryError};
 use std::collections::HashMap;
