+++ tests/reservoir_inertial_tests.rs
@@ -1,4 +1,8 @@
 //! Tests for InertialESN second-order momentum dynamics (ADR-0064).
+//!
+//! Float comparisons allowed: beta values are exact constants set by builder.
+
+#![allow(clippy::float_cmp)]
 
 use chaotic_semantic_memory::reservoir::{ChaoticReservoir, Reservoir};
